refactor(boards-list): split board list into banner and list subcomponents
