+++ 03-maincamp-triptalk/03-CNA/src/app/boards/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import BoardList from "@/components/board-list";
+import BoardList from "@/components/board-list/list";
 
 export default function BoardListPage() {
   return (
