+++ 03-maincamp-triptalk/03-CNA/src/components/board-list/list/hook.ts
