+++ 03-maincamp-triptalk/03-CNA/src/components/board-list/list/queries.ts
